refactoring API calls out to seperate API file
